backend/requirements.txt
Flask==2.2.5
psycopg2-binary
prometheus-client
